Tharusha

+++ backend/app/Http/Controllers/UserController.php
@@ -18,15 +18,16 @@ class UserController extends Controller
     }
 
     // Update: Update user details
-    public function update(Request $request, $id)
+    public function update(Request $request)
     {
-        $user = User::findOrFail($id);
-
+        $user = Auth::user();
         $validatedData = $request->validate([
             'full_name' => 'required|string|max:128',
-            'email' => 'required|email|unique:users,email,' . $id,
-            'username' => 'required|string|max:32|unique:users,username,' . $id,
-        ]);
+            'email' => 'required|email|unique:users,email,' . $request->user()->id,
+            'username' => 'required|string|max:32|unique:users,username,' . $request->user()->id,
+            'uni_reg_no' => 'required|string|max:10|unique:users,uni_reg_no,' . $request->user()->id,
+            'university_id' => 'required|exists:universities,id',
+    ]);
 
         $user->update($validatedData);
 
@@ -34,27 +35,32 @@ class UserController extends Controller
     }
 
     // Delete: Delete a user
-    public function destroy($id)
+    public function destroy(Request $request)
     {
-        $user = User::findOrFail($id);
+        $user = Auth::user();
+         if (!Hash::check($request->password, Auth::user()->password)) {
+            return response()->json(['message' => 'Password is incorrect'], 401);
+        }
         $user->delete();
 
         return response()->json(['message' => 'User deleted successfully']);
     }
 
     // Patch: Update password
-    public function patch(Request $request, $id)
+    public function patch(Request $request)
     {
-        $user = User::findOrFail($id);
-
-        $validatedData = $request->validate([
-            'password' => 'required|string|min:8|confirmed',
+        $validatedRequest = $request->validate([
+            'oldPassword' => 'required|string',
+            'newPassword' => 'required|min:6|max:16|confirmed',
         ]);
+        $user = Auth::user();
 
-        $user->password = Hash::make($validatedData['password']);
+        $newPassword = $validatedRequest['newPassword'];
+
+        $user->password = Hash::make($newPassword);
         $user->save();
 
-        return response()->json(['message' => 'Password updated successfully']);
+        return response()->json(['message' => 'Password updated'], 200);
     }
 }
